# fix(scheduler): remove the defer function cost